fix(schema): drop .int() — zod 4 converters add implicit safe-integer bounds
Anthropic + OpenAI structured-output modes reject the implicit
'minimum: -2^53, maximum: 2^53' that zod 4's converter emits for
.int(). Switching to z.number(); the system prompts already instruct
LLMs to return integer scores.

Affects: assessmentOutputSchema (overall_score + dimension score),
matchScoreOutputSchema (score), tailor editOp (index, bullet_index).

+++ src/lib/ai/schemas/assessment.ts
@@ -2,13 +2,15 @@ import { z } from 'zod';
 import { profileSchema } from './profile';
 
 // Schema notes for LLM-output compatibility (Anthropic + OpenAI strict modes):
-//   - No .min()/.max() on integers — providers reject `minimum`/`maximum` on type:integer
+//   - No .min()/.max() on integers — providers reject `minimum`/`maximum`
 //   - No .min()/.max() on arrays — providers reject `minItems`/`maxItems`
-//   - Constraints are described in the system prompt instead and enforced
-//     post-hoc with .refine() if needed.
+//   - No .int() — zod 4 converters add implicit JS-safe-integer bounds
+//     (`minimum: -2^53, maximum: 2^53`) which providers also reject. Use
+//     z.number() and instruct the LLM to return integers via the prompt.
+//   - Constraints are described in the system prompt instead.
 
 const dimensionAssessment = z.object({
-  score: z.number().int().nullable(),
+  score: z.number().nullable(),
   evidence: z.string(),
   gaps: z.array(z.string()),
   strengths: z.array(z.string()),
@@ -29,7 +31,7 @@ export const assessmentInputSchema = z.object({
 export type AssessmentInput = z.infer<typeof assessmentInputSchema>;
 
 export const assessmentOutputSchema = z.object({
-  overall_score: z.number().int(),
+  overall_score: z.number(),
   dimensions: z.record(z.string(), dimensionAssessment),
   candid_summary: z.string(),
   next_steps: z.array(nextStep),
